feat(helm): added jwt secret auto gen (#1062)
Introduced a new auto generated secret that will only be altered on
first install called "$(akhq.name)-jwt-secret". The value of this
secret is added via an environment variable setting to the depoyment
only if no user defined variable with the same name is found. This
should avoid duplicate env vars with the same name that can cause
trouble with subsequent changes to the deployment object.